+++ src/Transformer/NullTransformer.php
@@ -26,7 +26,7 @@ final class NullTransformer implements TransformerInterface
         mixed $source,
         mixed $target,
         Type $sourceType,
-        Type $targetType,
+        ?Type $targetType,
         array $context
     ): mixed {
         if ($target !== null) {