+++ src/Transformer/ObjectToObjectTransformer.php
@@ -56,9 +56,13 @@ final class ObjectToObjectTransformer implements TransformerInterface, MainTrans
         mixed $source,
         mixed $target,
         Type $sourceType,
-        Type $targetType,
+        ?Type $targetType,
         array $context
     ): mixed {
+        if ($targetType === null) {
+            throw new InvalidArgumentException('Target type must not be null.');
+        }
+
         // get object cache
 
         if (!isset($context[MainTransformer::OBJECT_CACHE])) {