+++ src/Transformer/ScalarToScalarTransformer.php
@@ -26,7 +26,7 @@ final class ScalarToScalarTransformer implements TransformerInterface
         mixed $source,
         mixed $target,
         Type $sourceType,
-        Type $targetType,
+        ?Type $targetType,
         array $context
     ): mixed {
         if (!is_scalar($source)) {