+++ tests/Fixtures/Transformer/MoneyToMoneyDtoTransformer.php
@@ -61,7 +61,7 @@ class MoneyToMoneyDtoTransformer implements TransformerInterface
         mixed $source,
         mixed $target,
         Type $sourceType,
-        Type $targetType,
+        ?Type $targetType,
         array $context
     ): mixed {
         if (